I did Score, end game, start, and timer.
Score needs some work, it needs to keep score. End game has a countdown timer and ends when it turns red numbers. Start, you need to go in scenes and MainMenu. Then press start and it will lead you to the game to play, BTW quit shows down in console and options button doesn't so anything.